Remove DWARFisms from Symtab::Module (#1575)
* Remove compilation directory from Module

This is a concept specific to DWARF. These functions are not documented.

* Remove DWARFisms from Symtab::Module

There is no need to store the CU DIE from which a Module instance is
derived. The address of the CU can be used to reconstitute the
entry in the .debug_info section using dwarf_addrdie.

Because Module.h is part of the public API for Dyninst, this also
removes the transitive dependency on libdw.